requirements.txt
Django==2.2.3
django-rest-auth==0.9.5
django-rest-framework==0.1.0
djangorestframework==3.10.1
psycopg2==2.8.3
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
